Refactor `<Minter>` with new types
